Add Dockerfile for oetoolkits
This image represents only the basic setup required for others to use
OE Toolkits. Needed because some tools require only the OE Toolkits to
work, example: https://github.com/OpenEye-Contrib/OEMicroservices/

Since this this the most basic setup, other images in this repo could
use it as the base image.